Deleting Database for App Versions <7, <1.1
App versions not on the 1.1+ track will stop working and be unable to pull an updated schedule. Please update your app, or use the offline JSON database selector to continue your use of app version 6.
All versions 7 and up, 1.1 and up, will be using the new database in my SHSS directory.